requirements.txt
streamlit
torch
sentence-transformers
transformers
faiss-cpu
numpy
pypdf
beautifulsoup4
requests
